requirements.txt
asgiref==3.7.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
